upd: Move CBlockUndo class from main
