+++ divi/src/BlockUndo.h
@@ -1,0 +1,41 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2014 The Bitcoin developers
+// Copyright (c) 2014-2015 The Dash developers
+// Copyright (c) 2015-2017 The PIVX Developers
+// Copyright (c) 2017-2020 The DIVI Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_BLOCKUNDO_H
+#define BITCOIN_BLOCKUNDO_H
+
+#ifndef BITCOIN_UNDO_H
+#include <undo.h>
+#endif
+
+#ifndef BITCOIN_SERIALIZE_H
+#include <serialize.h>
+#endif
+
+#ifndef BITCOIN_CHAIN_H
+#include <chain.h>
+#endif
+
+/** Undo information for a CBlock */
+class CBlockUndo
+{
+public:
+    std::vector<CTxUndo> vtxundo; // for all but the coinbase
+
+    ADD_SERIALIZE_METHODS;
+
+    template <typename Stream, typename Operation>
+    inline void SerializationOp(Stream& s, Operation ser_action, int nType, int nVersion)
+    {
+        READWRITE(vtxundo);
+    }
+
+    bool WriteToDisk(CDiskBlockPos& pos, const uint256& hashBlock);
+    bool ReadFromDisk(const CDiskBlockPos& pos, const uint256& hashBlock);
+};
+#endif
